Drop child process adapter
While fully non-blocking the child process adapter is a factor 100 slower than all the other adapters combined, including the blocking fallback adapter. As such removing it wil increase performance for everyone not having ext-uv or ext-eio (coming in a follow-up PR) installed. Even at the cost of blocking for a short period of time.